pyproject.toml
[project]
name = "fiction-pipeline"
version = "0.1.0"
description = "Fiction Writing Pipeline - schema validation, relationship tracking, and context management"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
